Restructured source with webpack.
