+++ .gitignore
@@ -2,3 +2,4 @@
 .idea
 node_modules
 .DS_Store
+dev-dist
