+++ package.json
@@ -1,9 +1,33 @@
 {
   "name": "cable-slider",
-  "version": "0.2.0",
+  "description": "",
+  "version": "0.1.21",
   "author": "Tristan Sach",
+  "license": "",
   "devDependencies": {
-
+    "@babel/cli": "^7.13.14",
+    "@babel/core": "^7.13.14",
+    "@babel/preset-env": "^7.13.12",
+    "babel-loader": "^8.2.2",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^5.2.0",
+    "html-webpack-plugin": "^5.3.1",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
+    "style-loader": "^2.0.0",
+    "terser-webpack-plugin": "^5.1.1",
+    "webpack": "^5.31.0",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3"
   },
-  "description": ""
+  "dependencies": {
+    "@babel/polyfill": "^7.12.1",
+    "jquery": "^3.5.1"
+  },
+  "scripts": {
+    "dev": "webpack --config webpack.dev.js",
+    "serve": "webpack --config webpack.dev.js && webpack serve --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js && webpack --config webpack.prod.js --env minimize=1"
+  }
 }
