+++ src/default-settings.js
@@ -1,0 +1,31 @@
+
+export const _default_settings = {
+    container: false,
+    next: false,
+    prev: false,
+    shown: 1,
+    orientation: 'horizontal',
+    align: 'left', // for multiple items, set the zero point
+    margin: false,
+    //loop: false,
+    continuous: false,
+    min_clones: 0,
+    draggable: true,
+    thumbs_container: false,
+    thumbs_next: false,
+    thumbs_prev: false,
+    thumbs_shown: 3,
+    thumbs_orientation: 'horizontal',
+    thumbs_align: 'center', // for multiple items, set the zero point
+    thumbs_margin: false,
+    //thumbs_loop: false,
+    thumbs_continuous: false,
+    thumbs_min_clones: 0,
+    auto_thumbs: false, // will remove existing html inside thumbs container if it exists
+    //bullets_container: false,
+    auto_play: false, // set a number for a time period. false or 0 will not auto play.
+    hover_pause: false,
+    auto_create: true,
+    active_height: true,
+    responsive: {}
+};
